update db migration, schema remove unneccesary files, models.. add role: super admin, admin, governance, public_user add middleware to verify admin routes

+++ app/Controllers/Http/AdminController.js
@@ -1,15 +1,18 @@
 'use strict'
 const AdminService = use('App/Services/AdminService');
-const AdminModel = use('App/Models/Admin');
+const UserModel = use('App/Models/User');
 const HelperUtils = use('App/Common/HelperUtils');
 const Const = use('App/Common/Const');
 const randomString = use('random-string');
 class AdminController {
 
-  async create({ request }) {
+  async create({ request, auth }) {
     try {
       const inputs = request.only(['wallet_address', 'role', 'firstname', 'lastname', 'email']);
-      // inputs.password = request.input('password');
+      const authRole = auth.user.role;
+      if (parseInt(authRole) < parseInt(inputs.role)) {
+        return HelperUtils.responseBadRequest('ERROR: you are now allowed to create this user!');
+      }
       console.log('Create Admin with params: ', inputs);
 
       const adminService = new AdminService();
@@ -20,9 +23,8 @@ class AdminController {
         return HelperUtils.responseBadRequest('Wallet is used');
       }
 
-      const admin = new AdminModel();
+      const admin = new UserModel();
       admin.fill(inputs);
-      // admin.signature = randomString(15);  // TODO: Fill any string
       admin.status = Const.USER_STATUS.ACTIVE;
       await admin.save();
 
@@ -40,39 +42,56 @@ class AdminController {
       return HelperUtils.responseErrorInternal('ERROR: create admin fail !');
     }
   }
-  async update({ request }) {
+  async update({ request, auth }) {
     try {
       const inputs = request.only(['role', 'firstname', 'lastname', 'email', 'wallet_address']);
-      const id = request.params.id
+      const id = request.params.id;
+      const authRole = auth.user.role;
+
       console.log('Update Admin with params: ', inputs);
       const adminService = new AdminService();
       const admin = await adminService.findUser({
-        id
+        id,
       });
+
       if (admin) {
+        // check if auth.user has role to modify this profile.
+        if (parseInt(admin.role) > parseInt(authRole)) {
+          return HelperUtils.responseBadRequest('Error: you are not allowed to modify this user!');
+        }
+        if (parseInt(inputs.role) > parseInt(authRole)) {
+          return HelperUtils.responseBadRequest('ERROR: you are trying to change role of this user to a higher lever than your role!');
+        }
+        // update.
         admin.merge(inputs)
         await admin.save();
         return HelperUtils.responseSuccess(admin);
       }
+      // user not exist.
       return HelperUtils.responseBadRequest('Error: user not exist!');
     } catch (e) {
       console.log(e);
       return HelperUtils.responseErrorInternal('ERROR: udpate admin fail!');
     }
   }
-  async delete({ request }) {
+  async delete({ request, auth }) {
     try {
-      const id = request.params.id
+      const id = request.params.id;
+      const authRole = auth.user.role;
+
       console.log('delete Admin with params: ', id);
       const adminService = new AdminService();
       const admin = await adminService.findUser({
         id
       });
-      if(admin){
-       await admin.delete()
-       return HelperUtils.responseSuccess(admin);
+      if (admin) {
+        if (parseInt(admin.role) > parseInt(authRole)) {
+          return HelperUtils.responseBadRequest("Error: you cannot delete this user!");
+        }
+        await admin.delete()
+        return HelperUtils.responseSuccess(admin);
       }
-      return HelperUtils.responseBadRequest("Error: Delete unexisting user!");
+      return HelperUtils.responseBadRequest("Error: Delete non-existing user!");
     } catch (e) {
       console.log(e);
       return HelperUtils.responseErrorInternal('ERROR: delete admin fail!');
@@ -98,13 +117,13 @@ class AdminController {
     }
   }
 
-  async adminDetail({params }) {
+  async adminDetail({ params }) {
     try {
       const id = params.id;
       const adminService = new AdminService();
-      const admins = await adminService.findUser({ id });
-      if(admins){
-        return HelperUtils.responseSuccess(admins);
+      const admin = await adminService.findUser({ id });
+      if (admin) {
+        return HelperUtils.responseSuccess(admin);
       }
       return HelperUtils.responseErrorInternal('ERROR: get admin detail fail !');
     } catch (e) {
